fix: globalnotfound added, fix sidebar ui, and added drawer for voice transaction
