+++ README.md
@@ -1,60 +1,56 @@
-# Hi, I'm Mike Arthur
+# TRACKLE
 
-Aspiring Software Developer | Always learning, experimenting, and improving my coding skills
-
----
-
-## 🛠 Skills & Tools
-
-**Languages:**  
-![Java](https://img.shields.io/badge/-Java-F89820?style=flat-square&logo=java&logoColor=white)
-![PHP](https://img.shields.io/badge/-PHP-777BB4?style=flat-square&logo=php&logoColor=white)
-![C](https://img.shields.io/badge/-C-A8B9CC?style=flat-square&logo=c&logoColor=white)
-![Python](https://img.shields.io/badge/-Python-3776AB?style=flat-square&logo=python&logoColor=white)
-![JavaScript](https://img.shields.io/badge/-JavaScript-F7DF1E?style=flat-square&logo=javascript&logoColor=black)
-![TypeScript](https://img.shields.io/badge/-TypeScript-3178C6?style=flat-square&logo=typescript&logoColor=white)
-
-**Databases:**  
-![MySQL](https://img.shields.io/badge/-MySQL-4479A1?style=flat-square&logo=mysql&logoColor=white)
-![PostgreSQL](https://img.shields.io/badge/-PostgreSQL-336791?style=flat-square&logo=postgresql&logoColor=white)
-
-**Frameworks:**  
-![Tailwind CSS](https://img.shields.io/badge/-Tailwind_CSS-06B6D4?style=flat-square&logo=tailwindcss&logoColor=white)
-![Bootstrap](https://img.shields.io/badge/-Bootstrap-7952B3?style=flat-square&logo=bootstrap&logoColor=white)
-![Laravel](https://img.shields.io/badge/-Laravel-FF2D20?style=flat-square&logo=laravel&logoColor=white)
-![Next.js](https://img.shields.io/badge/-Next.js-000000?style=flat-square&logo=nextdotjs&logoColor=white)
-
-**Tools & Platforms:**  
-![Docker](https://img.shields.io/badge/-Docker-2496ED?style=flat-square&logo=docker&logoColor=white)
-![Supabase](https://img.shields.io/badge/-Supabase-3ECF8E?style=flat-square&logo=supabase&logoColor=white)
-![Postman](https://img.shields.io/badge/-Postman-FF6C37?style=flat-square&logo=postman&logoColor=white)
-![Figma](https://img.shields.io/badge/-Figma-F24E1E?style=flat-square&logo=figma&logoColor=white)
-![GitHub](https://img.shields.io/badge/-GitHub-181717?style=flat-square&logo=github&logoColor=white)
+**TRACKLE** is a smart personal finance tracking application that empowers users to stay on top of their money.  
+With AI-powered insights, clear visualizations, and intuitive tools, TRACKLE makes it easy to monitor income, expenses, and budgets—all in one place.
 
 ---
 
-## 🌱 Currently Learning
+## Features
 
-- AI-powered applications (Gemini AI, chatbots)
-- Cloud platform tools (deployment, DevOps basics)
+### 1. Dashboard
+
+The dashboard provides a **comprehensive overview of your financial health** at a glance.
+
+- **Total Balance**: Always know how much money you have.
+- **This Month’s Income & Expenses**: Track current financial activity in real time.
+- **AI-Generated Insights**: Get personalized recommendations, trends, and warnings powered by AI.
+- **Income vs Expense Comparison**: Visualize whether you’re earning more than you’re spending.
+- **Cash Flow Trend**: Identify spending patterns and predict future cash flow.
+- **Spending Breakdown**: See where your money goes with clear category-based charts.
+- **Budget Progress**: Monitor how close you are to reaching or exceeding your budget.
+
+### 2. Transactions
+
+Easily manage all your income and expense records.
+
+- **Add, Edit, Delete Transactions**: Keep your records accurate and up-to-date.
+- **Multiple Input Modes**:
+  - Manual entry for precise control
+  - **Speak to Save** — add transactions hands-free with voice input
+- **Smart Filters**: Quickly search and organize transactions by date, category, or type.
+
+### 3. Budget
+
+Take control of your spending by setting up tailored budgets.
+
+- **Budget Overview**: View total budget, amount spent, and remaining balance.
+- **Flexible Management**: Add, edit, or delete budgets as your financial goals evolve.
+- **Recurring Budgets**: Automate your monthly or weekly budgeting process.
+- **Carry Over**: Roll unused funds into the next cycle.
+- **Budget Filters**: Segment and analyze budgets across categories and timeframes.
+
+### 4. AI Assistant
+
+Your **personal finance chatbot** that brings intelligence to your data.
+
+- **Conversational Assistant**: Ask questions about your spending, income, or budgets.
+- **Data-Aware Responses**: The AI has access to your financial data, enabling contextual and personalized answers.
+- **Insightful Guidance**: Get suggestions on saving, budgeting, and improving cash flow.
 
 ---
 
-## 💻 Projects
+## Tech Stack
 
-- **MedNexus** – School clinic web app handling appointments and medical documents for students and employees.
-- **Chi-Square Calculator** – Statistical hypothesis test calculator.
-- **[SparkQuiz](https://sparkquiz.vercel.app/)** – MCQ generator using Gemini AI.
-- **Travelling Chatbot** – AI chatbot using Gemini to assist with travel questions and plans.
-- **[Trackle](https://trackle-web.vercel.app/)** – Financial app with AI insights and chatbot for tracking money (powered by Gemini, in progress).
-
----
-
-## 📫 Contact Me
-
-- **Email:** [mikearthurminoza@gmail.com](mailto:mikearthurminoza@gmail.com)
-- **Phone:** +63 991 190 3093
-
----
-
-> "The best way to predict the future is to invent it." – Alan Kay
+- **Frontend**: [Next.js](https://nextjs.org/) (React, TailwindCSS, Shadcn/UI)
+- **Backend**: [Supabase](https://supabase.com/) (PostgreSQL Database, Authentication, APIs)
+- **AI**: [Google Gemini](https://deepmind.google/technologies/gemini/) for insights and chatbot support
