+++ app/globals.css
@@ -22,7 +22,7 @@
     --accent-foreground: 343.4483 23.9669% 23.7255%;
     --destructive: 0 62.1891% 60.5882%;
     --destructive-foreground: 260 23.0769% 97.451%;
-    --border: 0 0% 92.1569%;
+    --border: 258.75 17.3913% 81.9608%;
     --input: 260 23.0769% 92.3529%;
     --ring: 260.4 22.9358% 57.2549%;
     --chart-1: 260.4 22.9358% 57.2549%;
