+++ components/sidebar/sidebar-header.tsx
@@ -1,4 +1,5 @@
 import Image from "next/image";
+import Link from "next/link";
 
 interface SidebarHeaderProps {
   isCollapsed?: boolean;
@@ -8,7 +9,7 @@ interface SidebarHeaderProps {
 export function SidebarHeader({ isCollapsed, isMobileOpen }: SidebarHeaderProps) {
   return (
     <div className="flex h-16 items-center border-b border-sidebar-border px-4">
-      <div className="flex items-center gap-1">
+      <Link href="/home" className="flex items-center gap-1">
         {/* Icon */}
         <div className="flex h-8 w-8 items-center justify-center rounded-lg bg-sidebar-primary">
           <Image src="/images/logo.png" alt="logo" width={32} height={32} />
@@ -19,7 +20,7 @@ export function SidebarHeader({ isCollapsed, isMobileOpen }: SidebarHeaderProps)
         >
           {(!isCollapsed || isMobileOpen) && "Trackle"}
         </span>
-      </div>
+      </Link>
     </div>
   );
 }
