+++ components/sidebar/sidebar-quick-actions.tsx
@@ -1,19 +1,29 @@
 "use client";
 
-import Link from "next/link";
-import { DollarSign, TrendingUp } from "lucide-react";
+import { useState } from "react";
+import { Download, PlusCircle, WalletMinimal } from "lucide-react";
 import { Button } from "@/components/ui/button";
+import TransactionVoiceDialog from "../transactions/TransactionVoiceDialog";
 
 const quickActions = [
   {
-    title: "Add Income",
-    href: "/transactions/add?type=income",
-    icon: TrendingUp,
+    title: "Add Transaction",
+    icon: PlusCircle,
+    onClick: (setDialogOpen: (open: boolean) => void) => setDialogOpen(true),
   },
   {
-    title: "Add Expense",
-    href: "/transactions/add?type=expense",
-    icon: DollarSign,
+    title: "Add Budget",
+    icon: WalletMinimal,
+    onClick: () => {
+      console.log("Open add budget dialog");
+    },
+  },
+  {
+    title: "Export Data",
+    icon: Download,
+    onClick: () => {
+      console.log("Exporting data");
+    },
   },
 ];
 
@@ -23,6 +33,8 @@ interface SidebarQuickActionsProps {
 }
 
 export function SidebarQuickActions({ isCollapsed, isMobileOpen }: SidebarQuickActionsProps) {
+  const [voiceDialogOpen, setVoiceDialogOpen] = useState(false);
+
   return (
     <div className="space-y-1">
       {!isCollapsed && (
@@ -32,23 +44,29 @@ export function SidebarQuickActions({ isCollapsed, isMobileOpen }: SidebarQuickA
           </h2>
         </div>
       )}
+
       {quickActions.map((action) => {
         const Icon = action.icon;
-
         return (
-          <Button
-            key={action.href}
-            variant="ghost"
-            className="w-full justify-start gap-3 px-3 py-2 h-10 text-sidebar-foreground hover:bg-sidebar-accent hover:text-sidebar-accent-foreground"
-            asChild
-          >
-            <Link href={action.href}>
+          <div key={action.title}>
+            <Button
+              variant="ghost"
+              className="w-full justify-start gap-3 px-3 py-2 h-10 text-sidebar-foreground hover:bg-sidebar-accent hover:text-sidebar-accent-foreground"
+              onClick={() => action.onClick(setVoiceDialogOpen)}
+            >
               <Icon className="h-4 w-4" />
               {(!isCollapsed || isMobileOpen) && action.title}
-            </Link>
-          </Button>
+            </Button>
+          </div>
         );
       })}
+
+      {/* Voice transaction dialog */}
+      <TransactionVoiceDialog
+        open={voiceDialogOpen}
+        onOpenChange={setVoiceDialogOpen}
+        onSubmit={() => {}}
+      />
     </div>
   );
 }
