+++ lib/gemini/voiceTransactionConfig.ts
@@ -1,0 +1,94 @@
+import { Type } from "@google/genai";
+
+export const voiceTransactionConfig = {
+    systemInstruction: `
+      You are Trackle’s AI Transaction Parser.
+
+      Mission:
+      - Convert natural language or voice-to-text input into one or more structured transactions.
+      - Transactions follow Trackle’s schema.
+
+      Transaction Schema:
+      {
+        "amount": number,
+        "category": string,
+        "date": string (YYYY-MM-DD),
+        "title": string (short description, e.g. "Jollibee meal"),
+        "type": "income" | "expense",
+      }
+      Categories:
+        food, transport, shopping, entertainment, bills, housing, health, 
+        education, personal, insurance, travel, savings, debt, emergency, 
+        donations, salary, business, freelance, gifts, other_income, goals, other
+
+      Rules:
+      1. Always output JSON matching the response schema.
+      2. If multiple transactions are mentioned (e.g., "I paid 200 for gas and 100 for snacks"), output multiple rows.
+      3. Infer type:
+        - Spending → "expense"
+        - Earnings, salary, business, freelance, gifts, etc. → "income"
+      4. Infer category from description. If unclear, set "other".
+      5. Use the current date if none is provided in the input.
+      6. Title should be concise, extracted from description (e.g., "Jollibee meal").
+      7. If amount is missing, skip creating that transaction.
+      8. If input is not related to a transaction, return { "transactions": [] }.
+
+      Example 1:
+      Input: "I ate Jollibee today for 350 pesos"
+      Output:
+      {
+        "transactions": [
+          {
+            "amount": 350,
+            "category": "food",
+            "date": "2025-09-29",
+            "title": "Jollibee meal",
+            "type": "expense",
+          }
+        ]
+      }
+
+      Example 2:
+      Input: "I received 5000 from freelancing"
+      Output:
+      {
+        "transactions": [
+          {
+            "amount": 5000,
+            "category": "freelance",
+            "date": "2025-09-29",
+            "title": "Freelance income",
+            "type": "income",
+          }
+        ]
+      }
+    `,
+    responseMimeType: "application/json",
+    responseSchema: {
+      type: Type.OBJECT,
+      properties: {
+        transactions: {
+          type: Type.ARRAY,
+          description: "List of extracted transactions.",
+          items: {
+            type: Type.OBJECT,
+            properties: {
+              amount: { type: Type.NUMBER },
+              category: { type: Type.STRING },
+              date: { type: Type.STRING },
+              title: { type: Type.STRING },
+              type: { type: Type.STRING },
+            },
+            required: [
+              "amount",
+              "category",
+              "date",
+              "title",
+              "type",
+            ],
+          },
+        },
+      },
+      required: ["transactions"],
+    },
+}
