+++ next.config.ts
@@ -1,8 +1,8 @@
 import type { NextConfig } from "next";
 
-const nextConfig: NextConfig = {
-  images: {
-    domains: ["assets.aceternity.com"],
+const nextConfig: NextConfig = { 
+  experimental: {
+    globalNotFound: true,
   },
 };
